refactor: 불필요한 변수 선언을 제거한다.
